r3223/sofa-dev : FIX: bug on 64bit when loading BMP Logo

+++ framework/sofa/helper/io/ImageBMP.cpp
@@ -51,7 +51,7 @@ bool ImageBMP::load(std::string filename)
         return false;
     }
     unsigned short int bfType;
-    long int bfOffBits;
+    uint32_t bfOffBits;
     short int biPlanes;
     short int biBitCount;
     long int biSizeImage;
@@ -79,7 +79,7 @@ bool ImageBMP::load(std::string filename)
     /* skip file size and reserved fields of bitmap file header */
     fseek(file, 8, SEEK_CUR);
     /* get the position of the actual bitmap data */
-    if (!fread(&bfOffBits, sizeof(long int), 1, file))
+    if (!fread(&bfOffBits, sizeof(uint32_t), 1, file))
     {
         std::cerr << "Error reading file!\n";
         return false;
